Refactor Controls component (#18)
* Refactor Controls component

* Update test language
