+++ src/components/Controls/Controls.test.tsx
@@ -1,144 +1,25 @@
 import React from 'react';
-import { receiveToken } from '../../store/main/main';
 import { shallow } from 'enzyme';
-import useLocalAudioToggle from '../../hooks/useLocalAudioToggle/useLocalAudioToggle';
-import useLocalVideoToggle from '../../hooks/useLocalVideoToggle/useLocalVideoToggle';
-import useRoomState from '../../hooks/useRoomState/useRoomState';
-import useScreenShareParticipant from '../../hooks/useScreenShareParticipant/useScreenShareParticipant';
-import useScreenShareToggle from '../../hooks/useScreenShareToggle/useScreenShareToggle';
 
 import Controls from './Controls';
+import useRoomState from '../../hooks/useRoomState/useRoomState';
 
-jest.mock('../../hooks/useLocalAudioToggle/useLocalAudioToggle');
-jest.mock('../../hooks/useLocalVideoToggle/useLocalVideoToggle');
 jest.mock('../../hooks/useRoomState/useRoomState');
-jest.mock('../../hooks/useScreenShareToggle/useScreenShareToggle');
-jest.mock('../../hooks/useScreenShareParticipant/useScreenShareParticipant');
-jest.mock('../../store/main/main');
-jest.mock('react-redux', () => ({ useDispatch: () => jest.fn() }));
 
-const mockUseLocalAudioToggle = useLocalAudioToggle as jest.Mock<any>;
-const mockUseLocalVideoToggle = useLocalVideoToggle as jest.Mock<any>;
 const mockUseRoomState = useRoomState as jest.Mock<any>;
-const mockUseScreenShareToggle = useScreenShareToggle as jest.Mock<any>;
-const mockUseScreenShareParticipant = useScreenShareParticipant as jest.Mock<any>;
-
-Object.defineProperty(navigator, 'mediaDevices', {
-  value: {
-    getDisplayMedia: () => {},
-  },
-  configurable: true,
-});
 
 describe('the Controls component', () => {
-  beforeEach(() => {
-    mockUseLocalAudioToggle.mockImplementation(() => [true, () => {}]);
-    mockUseLocalVideoToggle.mockImplementation(() => [true, () => {}]);
-    mockUseScreenShareToggle.mockImplementation(() => [false, () => {}]);
-  });
-
-  describe('End Call button', () => {
-    it('should not render when not connected to a room', () => {
-      const wrapper = shallow(<Controls />);
-      expect(wrapper.find('CallEndIcon').exists()).toBe(false);
-    });
-
-    it('should render when connected to a room', () => {
-      mockUseRoomState.mockImplementation(() => 'connected');
-      const wrapper = shallow(<Controls />);
-      expect(wrapper.find('CallEndIcon').exists()).toBe(true);
-    });
-
-    it('should delete the token from redux when clicked', () => {
-      mockUseRoomState.mockImplementation(() => 'connected');
-      const wrapper = shallow(<Controls />);
-      wrapper
-        .find('WithStyles(ForwardRef(Tooltip))')
-        .at(3)
-        .simulate('click');
-      expect(receiveToken).toHaveBeenCalledWith('');
-    });
-  });
-
-  it('should render correctly when audio is enabled', () => {
-    mockUseLocalAudioToggle.mockImplementation(() => [true, () => {}]);
+  it('should not render the ScreenShare and EndCall buttons when not connected to a room', () => {
+    mockUseRoomState.mockImplementation(() => 'disconnected');
     const wrapper = shallow(<Controls />);
-    expect(wrapper.find('MicIcon').exists()).toBe(true);
-    expect(wrapper.find('MicOffIcon').exists()).toBe(false);
-    expect(
-      wrapper
-        .find('WithStyles(ForwardRef(Tooltip))')
-        .at(0)
-        .prop('title')
-    ).toBe('Mute Audio');
+    expect(wrapper.find('ToggleScreenShareButton').exists()).toBe(false);
+    expect(wrapper.find('EndCallButton').exists()).toBe(false);
   });
 
-  it('should render correctly when audio is disabled', () => {
-    mockUseLocalAudioToggle.mockImplementation(() => [false, () => {}]);
+  it('should render the ScreenShare and EndCall buttons when connected to a room', () => {
+    mockUseRoomState.mockImplementation(() => 'connected');
     const wrapper = shallow(<Controls />);
-    expect(wrapper.find('MicIcon').exists()).toBe(false);
-    expect(wrapper.find('MicOffIcon').exists()).toBe(true);
-    expect(
-      wrapper
-        .find('WithStyles(ForwardRef(Tooltip))')
-        .at(0)
-        .prop('title')
-    ).toBe('Unmute Audio');
-  });
-
-  it('should render correctly when video is enabled', () => {
-    mockUseLocalVideoToggle.mockImplementation(() => [true, () => {}]);
-    const wrapper = shallow(<Controls />);
-    expect(wrapper.find('VideocamIcon').exists()).toBe(true);
-    expect(wrapper.find('VideocamOffIcon').exists()).toBe(false);
-    expect(
-      wrapper
-        .find('WithStyles(ForwardRef(Tooltip))')
-        .at(1)
-        .prop('title')
-    ).toBe('Mute Video');
-  });
-
-  it('should render correctly when video is disabled', () => {
-    mockUseLocalVideoToggle.mockImplementation(() => [false, () => {}]);
-    const wrapper = shallow(<Controls />);
-    expect(wrapper.find('VideocamIcon').exists()).toBe(false);
-    expect(wrapper.find('VideocamOffIcon').exists()).toBe(true);
-    expect(
-      wrapper
-        .find('WithStyles(ForwardRef(Tooltip))')
-        .at(1)
-        .prop('title')
-    ).toBe('Unmute Video');
-  });
-
-  it('should render correctly when screenSharing is allowed', () => {
-    const wrapper = shallow(<Controls />);
-    expect(wrapper.find('ScreenShareIcon').exists()).toBe(true);
-  });
-
-  it('should render correctly when the user is sharing their screen', () => {
-    mockUseScreenShareToggle.mockImplementation(() => [true, () => {}]);
-    const wrapper = shallow(<Controls />);
-    expect(wrapper.find('StopScreenShareIcon').exists()).toBe(true);
-  });
-
-  it('should render correctly when another user is sharing their screen', () => {
-    mockUseScreenShareParticipant.mockImplementation(() => 'mockParticipant');
-    mockUseScreenShareToggle.mockImplementation(() => [false, () => {}]);
-    const wrapper = shallow(<Controls />);
-    expect(
-      wrapper
-        .find('WithStyles(ForwardRef(Fab))')
-        .at(2)
-        .prop('disabled')
-    ).toBe(true);
-  });
-
-  it('should not render the screenshare button if screensharing is not supported', () => {
-    Object.defineProperty(navigator, 'mediaDevices', { value: { getDisplayMedia: undefined } });
-    const wrapper = shallow(<Controls />);
-    expect(wrapper.find('ScreenShareIcon').exists()).toBe(false);
-    expect(wrapper.find('StopScreenShareIcon').exists()).toBe(false);
+    expect(wrapper.find('ToggleScreenShareButton').exists()).toBe(true);
+    expect(wrapper.find('EndCallButton').exists()).toBe(true);
   });
 });
