+++ src/components/Controls/ToggleVideoButton/ToggleVideoButton.tsx
@@ -1,0 +1,34 @@
+import React from 'react';
+import { createStyles, makeStyles, Theme } from '@material-ui/core/styles';
+
+import Fab from '@material-ui/core/Fab';
+import Tooltip from '@material-ui/core/Tooltip';
+import Videocam from '@material-ui/icons/Videocam';
+import VideocamOff from '@material-ui/icons/VideocamOff';
+
+import useLocalVideoToggle from '../../../hooks/useLocalVideoToggle/useLocalVideoToggle';
+
+const useStyles = makeStyles((theme: Theme) =>
+  createStyles({
+    fab: {
+      margin: theme.spacing(1),
+    },
+  })
+);
+
+export default function ToggleVideoButton() {
+  const classes = useStyles();
+  const [isVideoEnabled, toggleVideoEnabled] = useLocalVideoToggle();
+
+  return (
+    <Tooltip
+      title={isVideoEnabled ? 'Mute Video' : 'Unmute Video'}
+      placement="top"
+      PopperProps={{ disablePortal: true }}
+    >
+      <Fab className={classes.fab} onClick={toggleVideoEnabled}>
+        {isVideoEnabled ? <Videocam /> : <VideocamOff />}
+      </Fab>
+    </Tooltip>
+  );
+}
